modularise the code and avoid DRY code
