+++ src/wallet/bdk_wallet/mod.rs
@@ -1,4 +1,4 @@
-mod error;
+pub mod error;
 mod keychain;
 mod wallet;
